fix: update deleteSession

+++ src/lib/session.ts
@@ -1,9 +1,28 @@
 'use server';
-import { SignJWT } from "jose";
+import { jwtVerify, SignJWT } from "jose";
 import { prisma } from "./prisma";
 import { cookies } from "next/headers";
 
 const SESSION_NAME = "session";
+
+const encrypt = async (sessionId: string) => {
+  const secret = new TextEncoder().encode(process.env.SESSION_SECRET);
+  const jwt = await new SignJWT({ sessionId })
+    .setProtectedHeader({ alg: "HS256" })
+    .setIssuedAt()
+    .setExpirationTime("2d")
+    .sign(secret);
+  return jwt;
+}
+
+export const decrypt = async (token: string) => {
+  const secret = new TextEncoder().encode(process.env.SESSION_SECRET);
+  const { payload } = await jwtVerify(token, secret, {
+    algorithms: ["HS256"],
+  });
+  
+  return payload.sessionId;
+}
 
 export const createSession = async (userId: string) => {
   const expiresAt = new Date(Date.now() + 2 * 24 * 60 * 60 * 1000)
@@ -15,8 +34,9 @@ export const createSession = async (userId: string) => {
   });
 
   const sessionId = sessionInsert.id;
+  const encryptedSession = await encrypt(sessionId);
   const cookieStore = await cookies();
-  cookieStore.set(SESSION_NAME, sessionId, {
+  cookieStore.set(SESSION_NAME, encryptedSession, {
     httpOnly: true,
     secure: process.env.NODE_ENV === "production",
     expires: expiresAt,
@@ -33,9 +53,12 @@ export const getSession = async () => {
   const session = cookieStore.get(SESSION_NAME)?.value;
   if (!session) return null;
 
+  const decryptedSession = await decrypt(session);
+  if (!decryptedSession) return null;
+
   // Get session from database
   const sessionData = await prisma.session.findUnique({
-    where: { id: session },
+    where: { id: decryptedSession.toString() },
     include: {
       user: {
         include: {
@@ -56,10 +79,16 @@ export const getSession = async () => {
 
 export const deleteSession = async () => {
   const cookieStore = await cookies();
-  const session = cookieStore.get(SESSION_NAME)?.value;
-  if (session) {
+  const session = cookieStore.get(SESSION_NAME)?.value!;
+  const decryptedSession = await decrypt(session);
+  if (!decryptedSession) {
+    console.log("Session not found");
+    return null
+  };
+
+  if (session && decryptedSession) {
     await prisma.session.delete({
-      where: { id: session},
+      where: { id: decryptedSession.toString() },
     })
   };
   cookieStore.delete(SESSION_NAME);
